# backend/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/v1/chat/chat/{guid}:
    post:
      consumes:
      - application/json
      description: 为指定项目添加新的对话消息
      parameters:
      - description: 项目GUID
        in: path
        name: guid
        required: true
        type: string
      - description: 对话消息
        in: body
        name: message
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 添加对话消息
      tags:
      - 对话消息
  /api/v1/chat/messages/{guid}:
    get:
      consumes:
      - application/json
      description: 获取指定项目的对话消息历史记录，支持分页
      parameters:
      - description: 项目GUID
        in: path
        name: guid
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 50
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 获取项目对话历史
      tags:
      - 对话消息
  /api/v1/chat/send-to-agent/{guid}:
    post:
      consumes:
      - application/json
      description: 用户向特定 Agent 发送消息，用于回答 Agent 的问题并恢复项目执行
      parameters:
      - description: 项目GUID
        in: path
        name: guid
        required: true
        type: string
      - description: 消息内容
        in: body
        name: message
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 向指定 Agent 发送消息
      tags:
      - 对话消息
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"